Add: unit tests and moved from main
