+++ assignment2/src/routes/mod.rs
@@ -1,0 +1,3 @@
+//! This module contains all the routes for the application.
+
+pub mod github_webhook;
